Introduce a QWidget based on QGraphicsView to load a selkie webapp
The separation is not quite nice yet, we need access to the toolbar and page,
I'll have to think a bit how we can factor this out nicely. With this patch,
we can make the graphicsview-based webview resize when the size of the window
changes.

With some glitches, this makes selkie on graphicsview work. :)